# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * DOR NER operational resilience API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type DashboardMode = typeof DashboardMode[keyof typeof DashboardMode];


export const DashboardMode = {
  baseline: 'baseline',
  disruption: 'disruption',
} as const;

export interface Kpis {
  accessibility: number;
  activeVehicles: number;
  deliveries: number;
  criticalDeliveries: number;
  riskCorridors: number;
  incidents: number;
  predictions: number;
  averageDelay: number;
}

export type MapStateNodeStatus = typeof MapStateNodeStatus[keyof typeof MapStateNodeStatus];


export const MapStateNodeStatus = {
  stable: 'stable',
  watch: 'watch',
  restricted: 'restricted',
  critical: 'critical',
} as const;

export interface MapStateNode {
  name: string;
  shortName: string;
  x: number;
  y: number;
  accessibility: number;
  status: MapStateNodeStatus;
}

export type CorridorStatus = typeof CorridorStatus[keyof typeof CorridorStatus];


export const CorridorStatus = {
  open: 'open',
  watch: 'watch',
  'high-risk': 'high-risk',
  blocked: 'blocked',
} as const;

export interface Corridor {
  id: string;
  name: string;
  points: number[][];
  status: CorridorStatus;
  accessibility: number;
  risk: number;
}

export interface MapVehicle {
  id: string;
  x: number;
  y: number;
  status: string;
  cargo: string;
}

export interface MapIncident {
  id: string;
  x: number;
  y: number;
  type: string;
  severity: string;
}

export interface MapHub {
  name: string;
  x: number;
  y: number;
  stockLevel: number;
}

export interface MapState {
  states: MapStateNode[];
  corridors: Corridor[];
  vehicles: MapVehicle[];
  incidents: MapIncident[];
  hubs: MapHub[];
}

export type PredictionLevel = typeof PredictionLevel[keyof typeof PredictionLevel];


export const PredictionLevel = {
  low: 'low',
  moderate: 'moderate',
  high: 'high',
  critical: 'critical',
} as const;

export interface Prediction {
  id: string;
  title: string;
  corridor: string;
  probability: number;
  confidence: number;
  window: string;
  factors: string[];
  impact: string;
  recommendation: string;
  alternative: string;
  level: PredictionLevel;
}

export interface SupplyRisk {
  district: string;
  item: string;
  daysRemaining: number;
  incomingEta: string;
  roadRisk: string;
  risk: string;
  recommendation: string;
}

export interface Activity {
  id: string;
  time: string;
  title: string;
  detail: string;
  tone: string;
}

export interface Impact {
  deliveriesProtected: number;
  vehiclesRerouted: number;
  riskAvoided: string;
  delayAvoided: string;
  responseTime: string;
}

export interface Dashboard {
  mode: DashboardMode;
  updatedAt: string;
  kpis: Kpis;
  map: MapState;
  predictions: Prediction[];
  supplyRisks: SupplyRisk[];
  activity: Activity[];
  impact: Impact;
}

export interface District {
  id: string;
  name: string;
  state: string;
  population: number;
  accessibility: number;
  daysOfSupply: number;
  activeIncidents: number;
  weather: string;
  risk: string;
  criticalNeeds: string[];
}

export interface Vehicle {
  id: string;
  type: string;
  cargo: string;
  capacity: number;
  driver: string;
  origin: string;
  destination: string;
  speed: number;
  eta: string;
  fuel: number;
  priority: string;
  risk: string;
  status: string;
}

export interface Delivery {
  id: string;
  cargo: string;
  quantity: number;
  origin: string;
  destination: string;
  eta: string;
  priority: string;
  status: string;
  risk: string;
}

export interface Incident {
  id: string;
  type: string;
  district: string;
  road: string;
  severity: string;
  status: string;
  timestamp: string;
  description: string;
}

export interface IncidentInput {
  type: string;
  district: string;
  road: string;
  severity: string;
  description: string;
}

export interface Alert {
  id: string;
  title: string;
  message: string;
  severity: string;
  status: string;
  createdAt: string;
  actions: string[];
}

export type AlertActionInputAction = typeof AlertActionInputAction[keyof typeof AlertActionInputAction];


export const AlertActionInputAction = {
  acknowledge: 'acknowledge',
  assign: 'assign',
  escalate: 'escalate',
  resolve: 'resolve',
} as const;

export interface AlertActionInput {
  action: AlertActionInputAction;
}

export interface RouteInput {
  origin: string;
  destination: string;
  cargo: string;
  quantity: number;
  priority: string;
  vehicleType: string;
}

export interface RouteOption {
  id: string;
  name: string;
  recommended: boolean;
  distance: string;
  eta: string;
  weatherRisk: number;
  landslideRisk: number;
  floodRisk: number;
  traffic: string;
  accessibility: number;
  fuel: string;
  overallRisk: string;
  score: number;
  explanation: string[];
}

export interface MissionInput {
  mission: string;
  destination: string;
  cargo: string;
  quantity: number;
  deadline: string;
  priority: string;
}

export interface Mission {
  id: string;
  mission: string;
  destination: string;
  vehicle: string;
  route: string;
  backupRoute: string;
  eta: string;
  risk: string;
  status: string;
}

export interface Analytics {
  deliverySuccess: number;
  averageDelay: number;
  routeReliability: number;
  shortages: number;
  incidents: number;
  utilization: number;
  accessibility: number;
  responseTime: number;
}

export interface SimulationResult {
  dashboard: Dashboard;
  chain: string[];
  headline: string;
}

export type GetAnalyticsParams = {
range?: GetAnalyticsRange;
};

export type GetAnalyticsRange = typeof GetAnalyticsRange[keyof typeof GetAnalyticsRange];


export const GetAnalyticsRange = {
  '7d': '7d',
  '30d': '30d',
  '90d': '90d',
} as const;