# lib/api-zod/src/generated/types/alert.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * DOR NER operational resilience API
 * OpenAPI spec version: 0.1.0
 */

export interface Alert {
  id: string;
  title: string;
  message: string;
  severity: string;
  status: string;
  createdAt: string;
  actions: string[];
}