# lib/api-zod/src/generated/types/analytics.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * DOR NER operational resilience API
 * OpenAPI spec version: 0.1.0
 */

export interface Analytics {
  deliverySuccess: number;
  averageDelay: number;
  routeReliability: number;
  shortages: number;
  incidents: number;
  utilization: number;
  accessibility: number;
  responseTime: number;
}