# lib/api-zod/src/generated/types/dashboard.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * DOR NER operational resilience API
 * OpenAPI spec version: 0.1.0
 */
import type { Activity } from './activity';
import type { DashboardMode } from './dashboardMode';
import type { Impact } from './impact';
import type { Kpis } from './kpis';
import type { MapState } from './mapState';
import type { Prediction } from './prediction';
import type { SupplyRisk } from './supplyRisk';

export interface Dashboard {
  mode: DashboardMode;
  updatedAt: string;
  kpis: Kpis;
  map: MapState;
  predictions: Prediction[];
  supplyRisks: SupplyRisk[];
  activity: Activity[];
  impact: Impact;
}