# lib/api-zod/src/generated/types/delivery.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * DOR NER operational resilience API
 * OpenAPI spec version: 0.1.0
 */

export interface Delivery {
  id: string;
  cargo: string;
  quantity: number;
  origin: string;
  destination: string;
  eta: string;
  priority: string;
  status: string;
  risk: string;
}