# lib/api-zod/src/generated/types/incidentInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * DOR NER operational resilience API
 * OpenAPI spec version: 0.1.0
 */

export interface IncidentInput {
  type: string;
  district: string;
  road: string;
  severity: string;
  description: string;
}