# lib/api-zod/src/generated/types/kpis.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * DOR NER operational resilience API
 * OpenAPI spec version: 0.1.0
 */

export interface Kpis {
  accessibility: number;
  activeVehicles: number;
  deliveries: number;
  criticalDeliveries: number;
  riskCorridors: number;
  incidents: number;
  predictions: number;
  averageDelay: number;
}