# lib/api-zod/src/generated/types/mapState.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * DOR NER operational resilience API
 * OpenAPI spec version: 0.1.0
 */
import type { Corridor } from './corridor';
import type { MapHub } from './mapHub';
import type { MapIncident } from './mapIncident';
import type { MapStateNode } from './mapStateNode';
import type { MapVehicle } from './mapVehicle';

export interface MapState {
  states: MapStateNode[];
  corridors: Corridor[];
  vehicles: MapVehicle[];
  incidents: MapIncident[];
  hubs: MapHub[];
}