# lib/api-zod/src/generated/types/prediction.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * DOR NER operational resilience API
 * OpenAPI spec version: 0.1.0
 */
import type { PredictionLevel } from './predictionLevel';

export interface Prediction {
  id: string;
  title: string;
  corridor: string;
  probability: number;
  confidence: number;
  window: string;
  factors: string[];
  impact: string;
  recommendation: string;
  alternative: string;
  level: PredictionLevel;
}