# lib/api-zod/src/generated/types/routeInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * DOR NER operational resilience API
 * OpenAPI spec version: 0.1.0
 */

export interface RouteInput {
  origin: string;
  destination: string;
  cargo: string;
  quantity: number;
  priority: string;
  vehicleType: string;
}