# lib/api-zod/src/generated/types/routeOption.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * DOR NER operational resilience API
 * OpenAPI spec version: 0.1.0
 */

export interface RouteOption {
  id: string;
  name: string;
  recommended: boolean;
  distance: string;
  eta: string;
  weatherRisk: number;
  landslideRisk: number;
  floodRisk: number;
  traffic: string;
  accessibility: number;
  fuel: string;
  overallRisk: string;
  score: number;
  explanation: string[];
}