# lib/api-zod/src/generated/types/simulationResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * DOR NER operational resilience API
 * OpenAPI spec version: 0.1.0
 */
import type { Dashboard } from './dashboard';

export interface SimulationResult {
  dashboard: Dashboard;
  chain: string[];
  headline: string;
}