# lib/api-zod/src/generated/types/supplyRisk.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * DOR NER operational resilience API
 * OpenAPI spec version: 0.1.0
 */

export interface SupplyRisk {
  district: string;
  item: string;
  daysRemaining: number;
  incomingEta: string;
  roadRisk: string;
  risk: string;
  recommendation: string;
}