# lib/api-zod/src/generated/types/vehicle.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * DOR NER operational resilience API
 * OpenAPI spec version: 0.1.0
 */

export interface Vehicle {
  id: string;
  type: string;
  cargo: string;
  capacity: number;
  driver: string;
  origin: string;
  destination: string;
  speed: number;
  eta: string;
  fuel: number;
  priority: string;
  risk: string;
  status: string;
}